requirements.txt
Django>=4.2.0
django-crontab>=0.7.1
djangorestframework>=3.14.0
feedparser>=6.0.10
mysqlclient>=2.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0